2026.03.17: * changed: scripts/*.bat: `del /Q ""` case workaround

+++ scripts/tests/bench/bat/test__read/test__read__12_by_set_using_tmp_file_w_del.bat
@@ -19,7 +19,8 @@ for /L %%i in (1,1,1000) do (
   set "TEMP_FILE=%TEMP_DIR%\tmp-!RANDOM!-!RANDOM!.txt"
   echo !RANDOM!-!RANDOM! > "!TEMP_FILE!"
   set /P VALUE=<"!TEMP_FILE!"
-  del /F /Q /A:-D "!TEMP_FILE!" >nul 2>nul
+  rem CAUTION: must check on empty variable to avoid accidental `del /Q ""` case
+  if defined TEMP_FILE del /F /Q /A:-D "!TEMP_FILE!" >nul 2>nul
 )
 
 endlocal
