+++ scripts/tools/locks/read_file_to_var.bat
@@ -49,8 +49,8 @@ call "%%~dp0busy_wait.bat" 20
 goto REPEAT_READ_LOOP
 
 :EXIT
-rem cleanup files
-del /F /Q /A:-D "%__LOCK_FILE0%" >nul 2>nul
+rem cleanup files, check on empty variable to avoid accidental `del /Q ""` case
+if defined __LOCK_FILE0 del /F /Q /A:-D "%__LOCK_FILE0%" >nul 2>nul
 
 (
   endlocal
