+++ scripts/tools/std/callshift.bat
@@ -164,7 +164,8 @@ rem redirect command line into temporary file to print it correctly
 
 set "__STRING__=" & for /F "usebackq tokens=* delims="eol^= %%i in ("%CMDLINE_TEMP_FILE%") do set "__STRING__=%%i"
 
-del /F /Q /A:-D "%CMDLINE_TEMP_FILE%" >nul 2>nul
+rem CAUTION: must check on empty variable to avoid accidental `del /Q ""` case
+if defined CMDLINE_TEMP_FILE del /F /Q /A:-D "%CMDLINE_TEMP_FILE%" >nul 2>nul
 
 rem WORKAROUND:
 rem   In case if `echo` is turned off externally.
@@ -235,7 +236,8 @@ if %LOCK_ACQUIRE% EQU 0 (
   goto CALL_LOCK_LOOP
 )
 
-del /F /Q /A:-D "%FLAG_LOCK_FILE%" >nul 2>nul
+rem CAUTION: must check on empty variable to avoid accidental `del /Q ""` case
+if defined FLAG_LOCK_FILE del /F /Q /A:-D "%FLAG_LOCK_FILE%" >nul 2>nul
 
 exit /b %LAST_ERROR%
